backend/course_service/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
uvicorn==0.24.0
python-multipart==0.0.6
asyncpg==0.29.0
pydantic-settings==2.1.0
httpx==0.28.1
PyJWT==2.8.0
python-jose==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
databases[postgresql]==0.8.0 